Skip SynthID tests when the extra is absent
CI installs only the dev extra, so importing transformers in the test module
failed the whole run rather than skipping it. The pyproject already declared a
`synthid` marker for exactly this — it just was not applied.

Tests now carry the marker and two independent skip conditions, reported
separately so a skip says which precondition was missing: the fixtures, or the
optional stack. Verified in a clean venv without the extra: 93 pass, 28 skip.

Also annotates a float return in the fixture generator that mypy flagged as Any.

+++ tests/fixtures/text/generate.py
@@ -425,7 +425,7 @@ def mean_g(processor: SynthIDTextWatermarkLogitsProcessor, ids: list[int]) -> fl
     # g is 0/1 longs: sum exactly in integers, divide once. No float accumulation,
     # so the value is bit-identical on every run and every machine.
     total = int(g.sum().item())
-    count = g.shape[1] * g.shape[2]
+    count = int(g.shape[1] * g.shape[2])
     return total / count
